% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adducts.neg}
\alias{adducts.neg}
\title{A data frame of most typical negative adducts and their mz values}
\format{
An object of class \code{data.frame} with 15 rows and 5 columns.
}
\usage{
adducts.neg
}
\description{
A data frame of most typical negative adducts and their mz values
}
\keyword{datasets}
