% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adducts.pos}
\alias{adducts.pos}
\title{A data frame of most typical positive adducts and their mz values}
\format{
An object of class \code{data.frame} with 34 rows and 5 columns.
}
\usage{
adducts.pos
}
\description{
A data frame of most typical positive adducts and their mz values
}
\keyword{datasets}
