% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atoms}
\alias{atoms}
\title{The isotopic and molecular masses of atoms}
\format{
An object of class \code{data.frame} with 118 rows and 7 columns.
}
\usage{
atoms
}
\description{
The isotopic and molecular masses of atoms
}
\keyword{datasets}
