% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_adducts}
\alias{find_adducts}
\title{Find adducts for molecular mass}
\usage{
find_adducts(mass, mode = "pos")
}
\arguments{
\item{mass}{numeric. Molecular mass of the precursor}

\item{mode}{regular expression. Which polarization to use.}
}
\value{
data.frame. The m/z values calculated for different adducts
}
\description{
Find adducts for molecular mass
}
