% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_by_compound}
\alias{find_by_compound}
\title{Find features by annotated names}
\usage{
find_by_compound(
  data,
  name,
  column_to_search = ".*Metabolite.*name",
  ignore.case = TRUE
)
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{name}{regular expression. The compound name to search}

\item{column_to_search}{regular expression. The column containing annotations}

\item{ignore.case}{logical. Whether to ignore case in search}
}
\value{
a vector of strings. The feature IDs of the hits
}
\description{
Find features by annotated names
}
