% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_by_MS2peaks}
\alias{find_by_MS2peaks}
\title{Find features by presence of MS2 peaks}
\usage{
find_by_MS2peaks(
  data,
  peaks,
  operator = all,
  tol = 0.005,
  lim = 0.03,
  ms2_col = "MS.*MS.*spectrum"
)
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{peaks}{numeric or vector of numeric values. The m/z values of the MS2 peaks to search}

\item{operator}{c(all, any) operator to use with multiple given peaks. Defaults to all.}

\item{tol}{numeric. m/z tolerance for peak matching}

\item{lim}{numeric. Relative abundance cutoff of peaks}

\item{ms2_col}{regular expression. Name of the MS2 spectrum column}
}
\value{
a vector of hits (feature IDs)
}
\description{
Find features by presence of MS2 peaks
}
