% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_by_mw}
\alias{find_by_mw}
\title{Find features by molecular mass or molecular formula}
\usage{
find_by_mw(data, query, tol = 0.005, mz_col = ".*mz")
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{query}{numeric/string. Either the molecular mass or the molecular formula for adduct search}

\item{tol}{numeric. m/z tolerance for peak matching}

\item{mz_col}{regular expression. Name of the m/z containing column}
}
\value{
matrix. feature_IDs, matched_adducts and accuracy
}
\description{
Find features by molecular mass or molecular formula
}
