% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_by_mz}
\alias{find_by_mz}
\title{Find features by m/z value}
\usage{
find_by_mz(data, mz, tol = 0.005, mz_col = ".*mz")
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{mz}{numeric. The m/z value for peak matching}

\item{tol}{numeric. m/z tolerance for peak matching}

\item{mz_col}{regular expression. Name of the m/z containing column}
}
\value{
a vector of hits (feature IDs)
}
\description{
Find features by m/z value
}
