% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_common_mw}
\alias{find_common_mw}
\title{Find shared molecular masses for two molecular features}
\usage{
find_common_mw(
  data,
  var1,
  var2,
  mode1 = var1,
  mode2 = var2,
  tol = 0.005,
  use.mass2adduct = FALSE,
  mz_col = ".*mz"
)
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{var1}{character string. Name of the 1st feature to extract}

\item{var2}{character string. Name of the 2nd feature to extract}

\item{mode1}{regular expression. Optional. Polarization of 1st adduct. Extracts from feature name by default.}

\item{mode2}{regular expression. Optional. Polarization of 2nd adduct. Extracts from feature name by default.}

\item{tol}{numeric. Δ m/z tolerance for adducts}

\item{use.mass2adduct}{boolean. Whether to also get fragment data from the package mass2adduct}

\item{mz_col}{regular expression. Name of the column containing m/z values}
}
\value{
a data.frame. Suggested adducts for each feature, accuracy and calculated average molecular mass
}
\description{
Checks whether the two molecular features have a shared molecular masses within given m/z tolerance.
Additionally, can use mass2adduct to find molecular fragments between the two features.
}
