% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_mwfromformula}
\alias{find_mwfromformula}
\title{Calculate accurate mass from a molecular formula}
\usage{
find_mwfromformula(mf, monoisotopic = TRUE)
}
\arguments{
\item{mf}{character string. Molecular formula.}

\item{monoisotopic}{logical. Whether to return monoisotopic mass (default) or molar mass.}
}
\value{
numeric. The mass of the molecule. NA if any of the symbols are not recognized or contain non-stable elements.
}
\description{
Calculate accurate mass from a molecular formula
}
