% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_mws}
\alias{find_mws}
\title{Find molecular masses for adduct}
\usage{
find_mws(data, var, mode = var, mz_col = "Average.*Mz")
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{var}{character string. Name of the feature to extract}

\item{mode}{regular expression. Optional. Which polarization to use. Extracts from feature name by default.}

\item{mz_col}{character string. Name of the m/z column}
}
\value{
data.frame. The molecular masses derived from different assumed adducts
}
\description{
Find molecular masses for adduct
}
