% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_mws2}
\alias{find_mws2}
\title{Find molecular masses for adduct m/z}
\usage{
find_mws2(mz, mode = "pos")
}
\arguments{
\item{mz}{numeric. m/z of the precursor}

\item{mode}{regular expression. Which polarization to use.}
}
\value{
data.frame. The molecular masses derived from different assumed adducts
}
\description{
Find molecular masses for adduct m/z
}
