% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{get_cors}
\alias{get_cors}
\title{Get correlation matrix between selected features}
\usage{
get_cors(exprs, names, use = "complete.obs", ...)
}
\arguments{
\item{exprs}{a numeric data.frame of matrix. A dataframe of peak areas features\*samples (rows\*columns)}

\item{names}{vector of character strings. The names of the features (feature IDs)}

\item{...}{all other params are passed to \code{cor()}}
}
\value{
a square matrix of correlations
}
\description{
Get correlation matrix between selected features
}
