% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{get_isotope_spectrum}
\alias{get_isotope_spectrum}
\title{Get isotope spectrum from feature data}
\usage{
get_isotope_spectrum(
  data,
  var,
  lim = 0,
  mz_col = "Average.Mz",
  isot_col = "MS1.isotopic.spectrum",
  adduct_col = "Adduct.type"
)
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{var}{character string. The rowname of the feature (feature_ID) to extract}

\item{lim}{numeric. Relative abundance cutoff of isotope peaks}

\item{mz_col}{character string. Name of the m/z column}

\item{isot_col}{character string. Name of the isotope spectrm column}

\item{adduct_col}{character string. Name of the adduct column}
}
\value{
A list of 4 objects:
precursor = m/z value of the precursor ion
isotope_spectrum = dataframe of the isotope spectrum
adduct = type of adduct
feature = the variable name
}
\description{
Get isotope spectrum from feature data
}
