% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{get_spectrum}
\alias{get_spectrum}
\title{Get MS2 spectrum from feature data}
\usage{
get_spectrum(
  data,
  var,
  lim = 0.01,
  itlim = 0,
  mz_col = ".*Mz.*",
  ms2_col = ".*MS.*MS.*spectrum",
  adduct_col = ".*Adduct.*",
  isot_col = ".*isotopic.*spectrum"
)
}
\arguments{
\item{data}{a data.frame object. The feature data of the MS peaks}

\item{var}{character string. The rowname of the feature (feature_ID) to extract}

\item{lim}{numeric. Relative abundance cutoff of MS2 peaks}

\item{itlim}{numeric. Relative abundance cutoff of isotope peaks}

\item{mz_col}{regular expression. Name of the m/z column}

\item{ms2_col}{regular expression. Name of the MS2 spectrm column}

\item{adduct_col}{regular expression. Name of the adduct column}

\item{isot_col}{regular expression. Name of the isotope spectrm column}
}
\value{
A list of 5 objects:
precursor = m/z value of the precursor ion
spectrum = dataframe of the MS2 spectru,
adduct = type of adduct
isotope_spectrum = dataframe of the isotope spectrum
feature = the variable name
}
\description{
Get MS2 spectrum from feature data
}
