% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{plot_isotope}
\alias{plot_isotope}
\title{Plot isotope spectrum}
\usage{
plot_isotope(sp, alpha = 0.6, fill = "blue3", width = width, ...)
}
\arguments{
\item{sp}{list. The spectrum returned by \code{get_spectrum} or \code{get_isotope_spectrum}}

\item{alpha}{numeric. Transparency of bars}

\item{...}{all other args are passed to \code{geom_col}}
}
\value{
ggplot object. The mass spec plot.
}
\description{
Plot isotope spectrum
}
