% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{read_tables_from_MSDIALfile}
\alias{read_tables_from_MSDIALfile}
\title{Read data from an MS-Dial format excel file (feature data \* abundances \* sample data)}
\usage{
read_tables_from_MSDIALfile(file, feature_IDs_present = TRUE)
}
\arguments{
\item{file}{string. The name of the .xlsx-file to read.}

\item{feature_IDs_present}{logical. Whether the first column contains unique feature-IDs for the variables}
}
\value{
A list with 3 elements (data frames):
Expressions or peak integrals
Metadata for the features
Metadata for the samples
}
\description{
Read data from an MS-Dial format excel file (feature data \* abundances \* sample data)
}
