% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_spectrum_explore.R
\name{find_mws2}
\alias{find_mws2}
\title{Find molecular weights for adduct m/z}
\usage{
find_mws2(mz, mode)
}
\arguments{
\item{mz}{numeric. m/z of the precursor}

\item{mode}{character string. Which polarization to use.}
}
\value{
data.frame. The molecular weights derived from different assumed adducts
}
\description{
Find molecular weights for adduct m/z
}
